Fix Pal ID's and replace datatypes recursively in SDK

+++ decomp/SDK/include/dolphin/os/OSReset.h
@@ -33,7 +33,7 @@ struct OSResetFunctionInfo
 void          OSRegisterResetFunction(OSResetFunctionInfo* info);
 void          OSUnregisterResetFunction(OSResetFunctionInfo* info);
 void          OSResetSystem(int reset, u32 resetCode, BOOL forceMenu);
-unsigned long OSGetResetCode();
+u32 OSGetResetCode();
 
 #ifdef __cplusplus
 }
