+++ decomp/SDK/src/dolphin/os/OSMutex.c
@@ -202,7 +202,7 @@ OSWaitCond(struct OSCond* cond, struct OSMutex* mutex)
 
     if (mutex->thread == currentThread)
     {
-        long count = mutex->count;
+        s32 count = mutex->count;
         mutex->count = 0;
         DEQUEUE_MUTEX(mutex, &currentThread->queueMutex, link);
         mutex->thread = 0;
@@ -247,7 +247,7 @@ __OSCheckMutex(struct OSMutex* mutex)
 {
     struct OSThread*      thread;
     struct OSThreadQueue* queue;
-    long                  priority;
+    s32                  priority;
 
     priority = 0;
     queue = &mutex->queue;
